Added Path class

+++ src/main/java/nl/naturalis/common/Sizeable.java
@@ -1,8 +1,9 @@
 package nl.naturalis.common;
 
 /**
- * Defines objects that can meaningfully be said to have a certain size or length.
- * 
+ * Defines objects that can meaningfully be said to have a certain size or
+ * length.
+ *
  * @author Ayco Holleman
  *
  */
@@ -10,9 +11,9 @@ public interface Sizeable {
 
   /**
    * Returns the size of the object.
-   * 
+   *
    * @return
    */
-  long size();
+  int size();
 
 }
